I make some changes in ui controller

+++ app/controller/uicontroller/productcontroller.js
@@ -4,45 +4,41 @@ const fs = require('fs')
 
 class productuicontroller {
 
+    // Show add product form
+    async addproductGet(req, res) {
+        res.render('productview/addproduct', { user: req.user });
+    }
+
     // Handle POST product
-    async addproduct(req, res) {
-        if (req.method === 'POST') {
-            try {
-                const { p_name, p_color, p_size, p_description, brand_name, active } = req.body;
-                if (!p_name || !p_color || !p_size || !p_description || !brand_name) {
-                    req.flash('err', 'All fields are required')
-                    return res.redirect('/addproduct');
-                }
-
-                // Handle checkbox for p_color
-                const p_colors = Array.isArray(p_color) ? p_color : [p_color];
-                const p_sizes = Array.isArray(p_size) ? p_size : [p_size];
-
-                if (!['true', 'false'].includes(active)) {
-                    req.flash('err', 'Please active this product or not')
-                    return res.redirect('/addproduct');
-                }
-
-                const productData = {
-                    p_name: p_name.trim(),
-                    p_color: p_colors,
-                    p_size: p_sizes,
-                    p_description: p_description.trim(),
-                    image: req.file.path, // Image path for handling image
-                    brand_name: brand_name.trim(),
-                    active: active
-                };
-                const product = new Product(productData);
-                await product.save();
-                req.flash('sucess', 'Product sucessfully added')
-                return res.redirect('/product');
-            } catch (error) {
-                console.error('Error saving product:', error);
-                req.flash('err', 'Error adding product')
+    async addproductPost(req, res) {
+        try {
+            const { p_name, p_color, p_size, p_description, brand_name } = req.body;
+            if (!p_name || !p_color || !p_size || !p_description || !brand_name) {
+                req.flash('err', 'All fields are required')
                 return res.redirect('/addproduct');
             }
+
+            // Handle checkbox for p_color
+            const p_colors = Array.isArray(p_color) ? p_color : [p_color];
+            const p_sizes = Array.isArray(p_size) ? p_size : [p_size];
+
+            const productData = {
+                p_name: p_name.trim(),
+                p_color: p_colors,
+                p_size: p_sizes,
+                p_description: p_description.trim(),
+                image: req.file.path, // Image path for handling image
+                brand_name: brand_name.trim(),
+            };
+            const product = new Product(productData);
+            await product.save();
+            req.flash('sucess', 'Product sucessfully added')
+            return res.redirect('/product');
+        } catch (error) {
+            console.error('Error saving product:', error);
+            req.flash('err', 'Error adding product')
+            return res.redirect('/addproduct');
         }
-        res.render('productview/addproduct', { user: req.user });
     }
 
     // Handle GET product
@@ -60,11 +56,6 @@ class productuicontroller {
 
     // Handle GET single product 
     async singleproduct(req, res) {
-        // Only admin can add the product
-        // if (req.user.role !== 'admin') {
-        //     req.flash('err', 'Only admin can edit product')
-        //     return res.redirect('/product');
-        // }
         const id = req.params.id;
         try {
             const product = await Product.findById(id);
@@ -82,32 +73,30 @@ class productuicontroller {
     async updateproduct(req, res) {
         const id = req.params.id;
         // Deleting image from uploads folder start
-        const product = await Product.findById(id); // Find product by id
-        const imagePath = path.resolve(__dirname, '../../../', product.image);
-        if (fs.existsSync(imagePath)) {
-            fs.unlink(imagePath, (err) => {
-                if (err) {
-                    console.error('Error deleting image file:', err);
-                } else {
-                    console.log('Image file deleted successfully:', product.image);
-                }
-            });
-        } else {
-            console.log('File does not exist:', imagePath);
+        if (req.file) {
+            const product = await Product.findById(id); // Find product by id
+            const imagePath = path.resolve(__dirname, '../../../', product.image);
+            if (fs.existsSync(imagePath)) {
+                fs.unlink(imagePath, (err) => {
+                    if (err) {
+                        console.error('Error deleting image file:', err);
+                    } else {
+                        console.log('Image file deleted successfully:', product.image);
+                    }
+                });
+            } else {
+                console.log('File does not exist:', imagePath);
+            }
         }
         // Deleting image from uploads folder end
         try {
-            const { p_name, p_color, p_size, p_description, brand_name, active } = req.body;
+            const { p_name, p_color, p_size, p_description, brand_name } = req.body;
             if (!p_name || !p_color || !p_size || !p_description || !brand_name) {
                 return res.status(400).send('All fields are required.');
             }
 
             const p_colors = Array.isArray(p_color) ? p_color : [p_color];
             const p_sizes = Array.isArray(p_size) ? p_size : [p_size];
-
-            if (!['true', 'false'].includes(active)) {
-                return res.status(400).send('Active status is required');
-            }
 
             const existingProduct = await Product.findById(id);
             if (!existingProduct) {
@@ -120,7 +109,6 @@ class productuicontroller {
                 p_size: p_sizes,
                 p_description: p_description.trim(),
                 brand_name: brand_name.trim(),
-                active: active,
                 image: req.file ? req.file.path : existingProduct.image
             };
 
@@ -132,6 +120,25 @@ class productuicontroller {
         } catch (error) {
             console.error('Error updating product:', error);
             return res.status(500).send('Error updating product');
+        }
+    }
+
+    // Handle Toggle Active Blog
+    async toggleProductActive(req, res) {
+        try {
+            const productId = req.params.id;
+            const product = await Product.findById(productId);
+            if (!product) {
+                return res.status(404).send("Product not found");
+            }
+            product.active = !product.active;
+            await product.save();
+            req.flash('sucess', "Active status change sucessfully")
+            res.redirect('/product');
+        } catch (error) {
+            console.error(error);
+            req.flash('err', "Active status is not changed")
+            res.status(500).send("Error updating product status");
         }
     }
